utils: move message() and banner() helpers
Move message() and banner() helpers function in utils module in order to
lighten the Controller module and make these helpers available to other
modules.

Also add minimal tests to cover these functions specifically.
